Extract HTTPauth route to a handler
