+++ handlers/views/http_authentication.haml
